Fix intensity fluctuation variance normalisation and OxD clamp-bound test precision
'variance' method used N-1 (sample) normalisation via var(im,0,3), but every
acquired frame is the full population being measured, not a sample -- switched
to var(im,1,3) (population, /N). testOxdBoundedBetweenZeroAndOne compared
clamped single-precision OxD against exact double literals 0.999/0.001, which
don't round-trip losslessly through single -- compare against
double(single(...)) instead. Both were the 2 pre-existing failures noted but
left open in @5d6062d; 53/53 tests now pass.

+++ tests/testRedoxCalibrate.m
@@ -68,9 +68,12 @@ classdef testRedoxCalibrate < matlab.unittest.TestCase
             ratioIn = tc.gradientRatio(4, 32, 3, 0.05, 1.1);  % deliberately outside [rMin rMax]
             p = tc.defaultParams();
             [oxd, ~] = redoxCalibrate(ratioIn, p);
-            tc.verifyGreaterThanOrEqual(double(min(oxd(:))), 0.001, ...
+            % Compare against the single-precision representations of the
+            % clamp bounds -- oxd is single, and single(0.999) round-trips
+            % to a value fractionally above the double literal 0.999.
+            tc.verifyGreaterThanOrEqual(double(min(oxd(:))), double(single(0.001)), ...
                 'OxD should be ≥ 0.001 (clamped).');
-            tc.verifyLessThanOrEqual(double(max(oxd(:))), 0.999, ...
+            tc.verifyLessThanOrEqual(double(max(oxd(:))), double(single(0.999)), ...
                 'OxD should be ≤ 0.999 (clamped).');
         end
